backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Pydantic
pydantic>=2.7.0
pydantic-settings>=2.3.0

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
pypdf==3.17.4

# LLM and Embeddings
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.7.2
anthropic==0.8.1

# Vector Store
faiss-cpu==1.7.4
chromadb==0.4.22

# Embeddings
sentence-transformers==2.2.2

# Task Queue
celery==5.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
numpy>=1.26.4
pandas==2.1.4
numpy-financial==1.0.0

# HTTP and CORS
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
